+++ jest.cfg.unit.js
@@ -4,7 +4,7 @@ module.exports = {
   setupFiles: [
     "./jest.setup.js"
   ],
-  testRegex: "src/__tests__/.+\\.spec\\.jsx?$",
+  testRegex: "(src|app)/__tests__/.+\\.spec\\.jsx?$",
   moduleFileExtensions: [
     "js",
     "jsx"